433MHz OOK Arduino Rx, usung Manchester protocol
This is a system to send and receive small packets of information over a 433MHz one way link.  Around 8-12 bytes of information is ideal and if people extend the above code multiple transmitters can be logged with one receiver as each transmitter can be identified within the packet and transmit it its own particular payload of data.